update login components

+++ app/web-authn.server.ts
@@ -43,7 +43,9 @@ export class WebAuthnService {
       const authenticationOptions = await generateAuthenticationOptions({
         rpID: this.#relyingPartyId,
 
-        allowCredentials: (await getPasskeysByUser(email)).map((passkey) => ({
+        allowCredentials: (
+          await getPasskeysByUser(email)
+        ).map((passkey) => ({
           id: passkey.credentialId,
           transports: passkey.transports,
         })),
@@ -129,9 +131,13 @@ export class WebAuthnService {
       where: { email },
       select: { challenge: true },
     });
-    await prisma.webAuthnChallenge.delete({
-      where: { email },
-    });
+
+    if (result?.challenge) {
+      await prisma.webAuthnChallenge.delete({
+        where: { email },
+      });
+    }
+
     return result?.challenge ?? null;
   };
